refactor: JwtTokenProvider 패키지 이동
- 해당 인터페이스는 Service 패키지보다는 Domain 패키지가 더 타당하다고 판단하여 이동
